common: move CheckedStaticBox to the main project
