+++ pcsx2/gui/CheckedStaticBox.cpp
@@ -13,7 +13,8 @@
  *  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "common/CheckedStaticBox.h"
+#include "PrecompiledHeader.h"
+#include "gui/CheckedStaticBox.h"
 
 CheckedStaticBox::CheckedStaticBox(wxWindow* parent, int orientation, const wxString& title)
 	: wxPanelWithHelpers(parent, wxVERTICAL)
